@fkleuver fixed nasty bug in experimental code
I experimented on something and there was a punctuator bug in the lexer that no tests for the lexer can catch :( I discovered it when playing with binary expr parsing.  The bug only surface if you repeat the code multiple times in a loop. It freeze because `state.currentChar` isn't set. We only advanced at the end of each punctuator